backend/requirements.txt
# Core Framework
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.6,<0.31.0
# pydantic: keep compatible with FastAPI (<3.0.0)
pydantic>=2.6.1,<3.0.0

# AI/ML Libraries
# Use the current Gemini SDK (matches google-genai already on your system)
google-genai>=1.46.0,<2.0.0
# Keep these pinned to avoid global env conflicts (google-genai requires them)
anyio>=4.8.0,<5.0.0
httpx>=0.28.1,<1.0.0
# Fix protobuf conflicts (clarifai-grpc/opentelemetry-proto require protobuf>=5)
protobuf>=5.29.5,<7.0.0

# Document Processing
PyPDF2>=3.0.1,<4.0.0
pypdf>=3.0.0,<7.0.0
python-pptx>=0.6.23,<1.0.0
python-docx>=1.1.0,<2.0.0
reportlab>=4.0.0,<5.0.0

# Database
# sqlalchemy: keep >=2.0.27 for stable SQLAlchemy 2.x
sqlalchemy>=2.0.27,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
# Alembic used for migrations (optional but safe)
alembic>=1.13.1,<2.0.0

# HTTP & API
requests>=2.31.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.6,<1.0.0

# Attendance Management (Excel processing)
pandas>=2.2.2,<3.0.0
openpyxl>=3.1.5,<4.0.0
